+++ sdk/compute/arm-compute/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
     
+## 21.3.0 (2023-10-26)
+    
+**Features**
+
+  - Added operation DedicatedHosts.beginRedeploy
+  - Added operation DedicatedHosts.beginRedeployAndWait
+  - Added Interface DedicatedHostsRedeployOptionalParams
+    
+    
 ## 21.2.0 (2023-08-14)
     
 **Features**
@@ -515,4 +524,4 @@ To understand the detail of the change, please refer to [Changelog](https://aka.
 
 To migrate the existing applications to the latest version, please refer to [Migration Guide](https://aka.ms/js-track2-migration-guide).
 
-To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+To learn more, please refer to our documentation [Quick Start](https://aka.ms/js-track2-quickstart).